✨ feat: add comebackNavigate, fix imports
