+++ src/features/onboarding/index.ts
@@ -9,8 +9,7 @@ import { and, debug, not, or } from "patronum";
 
 import { api } from "~/shared/api";
 import { createFlag } from "~/shared/lib/localstorage-flags";
-import { routes } from "~/shared/routing";
-import { comebackSave } from "~/shared/routing/comeback";
+import { comebackSave, routes } from "~/shared/routing";
 import { $viewer } from "~/shared/viewer";
 
 /**
