+++ src/shared/routing/index.ts
@@ -1,23 +1,3 @@
-import { createHistoryRouter, createRouterControls } from "atomic-router";
-import { sample } from "effector";
-import { createBrowserHistory } from "history";
-
-import { appStarted } from "~/shared/init";
-
-import { pageNotFoundRoute, routesMap } from "./routes";
-
 export { routes, pageNotFoundRoute } from "./routes";
-
-export const controls = createRouterControls();
-
-export const router = createHistoryRouter({
-  routes: routesMap,
-  controls,
-  notFoundRoute: pageNotFoundRoute,
-});
-
-sample({
-  clock: appStarted,
-  fn: () => createBrowserHistory(),
-  target: router.setHistory,
-});
+export { controls, router } from "./router";
+export { comebackNavigate, comebackRestore, comebackSave } from "./comeback";
